preflight runner extension parity
